refactor: package functions renamed to canisters (#1267)
