+++ studio/src/app/canisters/slides/slides.did
