+++ studio/src/app/canisters/slides/slides.utils.did.js
